attempt to port to emscripten

+++ src/scr/infcubes.c
@@ -152,7 +152,7 @@ static int gen_phong_tex(struct pimage *img, int xsz, int ysz, float sexp,
 		for(j=0; j<xsz; j++) {
 			float d = sqrt(u * u + v * v);
 			float val = pow(cos(d * M_PI / 2.0f), sexp);
-			int ival = abs(val * 255.0f);
+			int ival = abs((int)(val * 255.0f));
 
 			int r = dr + ival * sr / 256;
 			int g = dg + ival * sg / 256;
